figured out issues with fragment tests
